scripts/requirements.txt
pycryptodome~=3.21.0
requests~=2.32.3